Improve incremental build support
Update generated files names format to be stable to avoid rebuilds, moving from indices to hashed details.
