+++ src/SourceGenerators/Uno.UI.SourceGenerators/XamlGenerator/XamlFileDefinition.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
+using Uno.UI.SourceGenerators.Helpers;
 using Uno.UI.SourceGenerators.XamlGenerator.XamlRedirection;
 
 namespace Uno.UI.SourceGenerators.XamlGenerator
@@ -18,7 +19,7 @@ namespace Uno.UI.SourceGenerators.XamlGenerator
 			Objects = new List<XamlObjectDefinition>();
 			FilePath = file;
 
-			UniqueID = SanitizedFileName + "_" + GetMd5Hash(FilePath);
+			UniqueID = SanitizedFileName + "_" + HashBuilder.Build(FilePath);
 		}
 
 		private string SanitizedFileName => Path
@@ -32,28 +33,5 @@ namespace Uno.UI.SourceGenerators.XamlGenerator
 		public string FilePath { get; private set; }
 
 		public string UniqueID { get; }
-
-		static string GetMd5Hash(string input)
-		{
-			using (MD5 md5Hash = MD5.Create())
-			{
-				// Convert the input string to a byte array and compute the hash.
-				var data = md5Hash.ComputeHash(Encoding.UTF8.GetBytes(input));
-
-				// Create a new Stringbuilder to collect the bytes
-				// and create a string.
-				var sBuilder = new StringBuilder();
-
-				// Loop through each byte of the hashed data 
-				// and format each one as a hexadecimal string.
-				for (int i = 0; i < data.Length; i++)
-				{
-					sBuilder.Append(data[i].ToString("x2", CultureInfo.InvariantCulture));
-				}
-
-				// Return the hexadecimal string.
-				return sBuilder.ToString();
-			}
-		}
 	}
 }
